Set a hard limit on the maximum file size
Node/V8 cannot handle strings longer than 256mb and attempts to determine the length of a buffer in excess of this value will throw an error.

Since this module is not intended for large files - where a streaming solution is more appropriate - then set a hard limit of 250mb for the file size, and truncate files larger than that.

The limit is set to 250mb rather than 256mb because busboy will truncate only after the limit is *passed* and so will potentially create buffers slightly longer than the maximum length. Setting to just under the limit allows for this to occur without crashing.