Allow support for files with no name
If a file is uploaded from a source other than an html form it is reasonable for it not to be named. Don't rely only on the name property existing to establish if a file is present. Instead assume that am empty buffer, with no name is not a file and that if either a name _or_ any data is present, then treat as a file.